update pakage expression

+++ app/src/androidTest/java/mathHandler/BasicMath.java
@@ -4,7 +4,7 @@ package mathHandler;
  * Created by Hung si da on 20/03/2016.
  */
 public class BasicMath {
-    public static float tong(float a,float b){
+    public static double tong(double a,double b){
         return a+b;
     }
